moved help dropdown to a component changed help to link to wiki
