+++ GUI/src/main/kotlin/ui/screens/DiffScreen.kt
@@ -3,6 +3,7 @@ package ui.screens
 import androidx.compose.foundation.focusable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.TopAppBar
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
@@ -36,6 +37,14 @@ fun DiffScreen(state: MutableState<AppState>) {
         // #####   Focus   #####
         LaunchedEffect(Unit) {
             focusRequester.requestFocus()
+        }
+
+        TopAppBar {
+            Row(modifier = Modifier.fillMaxWidth()) {
+                // #####   Save Collage Button   #####
+                helpMenu(Modifier.weight(0.1f))
+                Spacer(modifier = Modifier.weight(0.9f))
+            }
         }
 
         // #####   Titles   #####
